+++ test-parts/part-b-external_modules/sub-modules/ip-strings.js
@@ -2,6 +2,9 @@ const testIP = "192.168.1.1";
 const invalidIP = "192.168.1.2.3";
 const testPingAddress = 'https://www.google.com/';
 
-exports.testString = testIP;
-exports.invalidString = invalidIP;
-exports.pingAddress = testPingAddress;
+module.exports =
+{
+	testString: testIP,
+	invalidString: invalidIP,
+	pingAddress: testPingAddress
+};