+++ test-parts/part-b-external_modules/sub-modules/os-strings.js
@@ -1,8 +1,8 @@
 const supportedOperatingSystemArray = ['aix', 'darwin', 'freebsd', 'linux', 'mac', 'openbsd', 'sunos', 'win32'];
 
-function checkOperatingSystemSupported(o)
+function checkOperatingSystemSupported(osStr)
 {
-	var searchIndex = supportedOperatingSystemArray.indexOf(o);
+	var searchIndex = supportedOperatingSystemArray.indexOf(osStr);
 	var searchResult = false;
 	
 	if (searchIndex >= 0 && searchIndex < supportedOperatingSystemArray.length)
@@ -13,5 +13,8 @@ function checkOperatingSystemSupported(o)
 	return searchResult;
 }
 
-exports.checkOsSupported = checkOperatingSystemSupported;
-exports.usesArray = true;
+
+module.exports =
+{
+	checkOsSupported: checkOperatingSystemSupported
+};