+++ test-parts/part-b-external_modules/sub-modules/require-node-module.js
@@ -1,17 +1,1 @@
-function requireModuleSafely(mName)
-{
-	var moduleRes = null;
-	
-	try
-	{
-		moduleRes = require(mName);
-	}
-	catch(e)
-	{
-		moduleRes = null;
-	}
-	
-	return moduleRes;
-}
-
-exports.requireModuleSafe = requireModuleSafely;
+// EMPTY